Working on SEO functionality
Added a new json file and set up so that SEO content (keywords,
description etc) are loaded into a collection. Temporarily disabling
the message that appears when an unsupported browser uses this site.

Removing the print template and all events associated with it. Added
the ms-seo package to insert seo content to the page on router after
action.

Add json content for SEO data, matching it to the name of the route.

Removed auto publish package.